backend/requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
django-rest-knox==4.2.0
gunicorn==20.1.0
